+++ packages/plugins/type-distributor/src/headless/import/load-file.helper.ts
@@ -1,0 +1,14 @@
+import { ssdImportStore } from '@/headless/stores'
+
+export async function loadFromLocal() {
+	const file = ssdImportStore.fileInput?.files?.[0]
+	if (!file) throw new Error('No file selected')
+
+	const xmlDocument = new DOMParser().parseFromString(
+		await file.text(),
+		'application/xml'
+	)
+
+	if (ssdImportStore.fileInput) ssdImportStore.fileInput.value = ''
+	ssdImportStore.loadFromSSD(xmlDocument, file.name)
+}