+++ packages/plugins/type-distributor/src/ui/components/index.ts
@@ -1,0 +1,1 @@
+export { default as BayTypeDetails} from './bay-type-details.svelte'