+++ packages/plugins/type-distributor/src/ui/components/items/index.ts
@@ -1,0 +1,3 @@
+export { default as EquipmentItem } from './items.equipment.svelte'
+export { default as FunctionItem } from './items.function.svelte'
+